requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
fastapi==0.135.3
h11==0.16.0
idna==3.11
numpy==2.4.4
pandas==3.0.2
pydantic==2.13.1
pydantic_core==2.46.1
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
schedule==1.2.2
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
jinja2